Developer information

InfiniteVectors.ZTransform — Type

ZTransform{S<:BiInfiniteVector} <: Function

ZTransform is a wrapper type for the ztransform function.

InfiniteVectors.Integers — Type

struct InfiniteVectors.Integers <: OrdinalRange{Int,Int}

Struct representing all integers.

InfiniteVectors.BiInfiniteVector — Type

Any subtype of InfiniteVector is a bi-infinite list of values, with one value for each integer in Z.

They are not iterable

InfiniteVectors.hascompactsupport — Function

hascompactsupport(vec::InfiniteVector)

Are their a finite number of non-zero elements.

InfiniteVectors.nexteven — Function

The first even number greater than or equal to n.

InfiniteVectors.nextodd — Function

The first odd number greater than or equal to n.

InfiniteVectors.previouseven — Function

The last even number, smaller than or equal to n.

InfiniteVectors.sublength — Function

sublength(vec::InfiniteVector) = vec.subvec

The number of non-zero indices