Developer information

Developer information

InfiniteVectors.ZTransform — Type.

ZTransform{S<:BiInfiniteVector} <: Function

ZTransform is a wrapper type for the ztransform function.

source

InfiniteVectors.Integers — Type.

struct InfiniteVectors.Integers <: OrdinalRange{Int,Int}

Struct representing all integers.

source

InfiniteVectors.BiInfiniteVector — Type.

Any subtype of InfiniteVector is a bi-infinite list of values, with one value for each integer in Z.

They are not iterable

source

InfiniteVectors.hascompactsupport — Function.

hascompactsupport(vec::InfiniteVector)

Are their a finite number of non-zero elements.

source

InfiniteVectors.nexteven — Function.

The first even number greater than or equal to n.

source

InfiniteVectors.nextodd — Function.

The first odd number greater than or equal to n.

source

InfiniteVectors.previouseven — Function.

The last even number, smaller than or equal to n.

source

InfiniteVectors.sublength — Function.

sublength(vec::InfiniteVector) = vec.subvec

The number of non-zero indices

source